+++ core/src/test/scala/unit/kafka/security/auth/SimpleAclAuthorizerTest.scala
@@ -315,28 +315,26 @@ class SimpleAclAuthorizerTest extends ZooKeeperTestHarness {
     }
 
     // Alternate authorizer, Remove all acls that end in 0
-    val runnables = acls.map { acl =>
-        new Runnable {
-          override def run(): Unit = {
-            val aclId = acl.principal.getName.toInt
-            if(aclId % 2 == 0) {
-              simpleAclAuthorizer.addAcls(Set(acl), commonResource)
-            } else {
-              simpleAclAuthorizer2.addAcls(Set(acl), commonResource)
-            }
-            if(aclId % 10 == 0) {
-              simpleAclAuthorizer2.removeAcls(Set(acl), commonResource)
-            }
-          }
+    val concurrentFuctions = acls.map { acl =>
+      () => {
+        val aclId = acl.principal.getName.toInt
+        if(aclId % 2 == 0) {
+          simpleAclAuthorizer.addAcls(Set(acl), commonResource)
+        } else {
+          simpleAclAuthorizer2.addAcls(Set(acl), commonResource)
         }
-    }.toList
+        if(aclId % 10 == 0) {
+          simpleAclAuthorizer2.removeAcls(Set(acl), commonResource)
+        }
+      }
+    }
 
     val expectedAcls = acls.filter { acl =>
       val aclId = acl.principal.getName.toInt
       aclId % 10 != 0
     }.toSet
 
-    assertConcurrent("Should support many concurrent calls", runnables, 15)
+    TestUtils.assertConcurrent("Should support many concurrent calls", concurrentFuctions, 15000)
 
     TestUtils.waitAndVerifyAcls(expectedAcls, simpleAclAuthorizer, commonResource)
     TestUtils.waitAndVerifyAcls(expectedAcls, simpleAclAuthorizer2, commonResource)
@@ -358,47 +356,5 @@ class SimpleAclAuthorizerTest extends ZooKeeperTestHarness {
     TestUtils.waitAndVerifyAcls(acls, simpleAclAuthorizer, resource)
 
     acls
-  }
-
-  /**
-    * https://github.com/junit-team/junit/wiki/Multithreaded-code-and-concurrency
-    *
-    * To use this you pass in a Collection of Runnables that are your arrange\act\assert test on the SUT,
-    * they all run at the same time in the assertConcurrent method; the chances of triggering a multithreading code error,
-    * and thereby failing some assertion are greatly increased.
-    */
-  private def assertConcurrent(message: String, runnables: List[Runnable], maxTimeoutSeconds: Int) {
-    val numThreads = runnables.size
-    val exceptions = Collections.synchronizedList(new util.ArrayList[Throwable]())
-    val threadPool = Executors.newFixedThreadPool(numThreads)
-    try {
-      val allExecutorThreadsReady = new CountDownLatch(numThreads)
-      val afterInitBlocker = new CountDownLatch(1)
-      val allDone = new CountDownLatch(numThreads)
-      for (submittedTestRunnable <- runnables) {
-        threadPool.submit(new Runnable() {
-          def run() {
-            allExecutorThreadsReady.countDown()
-            try {
-              afterInitBlocker.await()
-              submittedTestRunnable.run()
-            } catch {
-              case e: Throwable => exceptions.add(e)
-            } finally {
-              allDone.countDown()
-            }
-          }
-        })
-      }
-      // wait until all threads are ready
-      assertTrue("Timeout initializing threads! Perform long lasting initializations before passing runnables to assertConcurrent",
-        allExecutorThreadsReady.await(runnables.size * 10, TimeUnit.MILLISECONDS))
-      // start all test runners
-      afterInitBlocker.countDown()
-      assertTrue(message +" timeout! More than" + maxTimeoutSeconds + "seconds", allDone.await(maxTimeoutSeconds, TimeUnit.SECONDS))
-    } finally {
-      threadPool.shutdownNow()
-    }
-    assertTrue(message + "failed with exception(s)" + exceptions, exceptions.isEmpty())
   }
 }
